feat: WidgetSnapshotUpdater 구현
